Number of btags should be an integer.

+++ MFVNeutralino/interface/MiniNtuple.h
@@ -42,7 +42,7 @@ namespace mfv {
     unsigned char jet_id[50];
     float jet_bdisc[50];
     float ht(float min_jet_pt=40.) const;
-    float nbtags(float min_bdisc) const;
+    int nbtags(float min_bdisc) const;
     float gen_x[2];
     float gen_y[2];
     float gen_z[2];
